chore: Refactor database.py and main.py
This commit refactors the code in `database.py` and `main.py` files. In `database.py`, the functions for fetching ticker URLs, fetching system information, and updating the system status have been moved to a separate module called `database`. In `main.py`, the import statements for the database functions have been updated to reflect the changes. This refactoring improves code organization and maintainability.

```
